Completed I2C block.
Completed I2C block, which will read and write a single byte of data to a slave.  If the slave doesn't ACK its address, the module goes into an idle state and generates an ACK error. If the module receives any invalid data during a read, it will request another byte until it either receives a valid byte or exceeds 5 retries.